feat: add pip-installable package
   - Create src/blinkg/ package with evaluate() function for programmatic use
   - Move evaluation scripts to evaluation/scripts/
   - Add pyproject.toml with package metadata
   - Organize examples into examples/lib/ and examples/hf/
   - Update README with library usage section
