+++ evaluation/scripts/clean_tables.py
